requirements.txt
requests>=2.31
python-dotenv>=1.0.1
pytest>=8.0.2